fix(ingestion): explicit $1::int cast in jsonb_build_array — fixes 'could not determine data type of parameter $1' that was aborting every raw message ingestion (0 listings created from 119 history messages).
When a parameter is used in both an INTEGER column context and a jsonb_build_array(any) context in the same query, Postgres can't reconcile the types. Forcing ::int resolves the inference.

Also: more verbose error logging in whatsapp.js to surface pg error codes + stack on future failures.

+++ server/services/exactDedup.js
@@ -186,8 +186,8 @@ async function findOrCreateCluster(msg, client = null) {
     if (isNewCluster) {
       await c.query(
         `UPDATE raw_clusters SET
-           representative_raw_message_id = $1,
-           all_raw_message_ids = jsonb_build_array($1)
+           representative_raw_message_id = $1::int,
+           all_raw_message_ids = jsonb_build_array($1::int)
          WHERE id = $2`,
         [rawMessageId, clusterId]
       );
